% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis.R
\name{apply_pro}
\alias{apply_pro}
\title{Apply Prognostic Model to New Data}
\usage{
apply_pro(trained_model_object, new_data, time_unit = "day")
}
\arguments{
\item{trained_model_object}{A trained object (class \code{pro_model}).}

\item{new_data}{Data frame of new patients.}

\item{time_unit}{Time unit for data preparation.}
}
\value{
Data frame with IDs, outcomes (if available), and risk scores.
}
\description{
Generates risk scores for new patients using a trained model.
}
