% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis.R
\name{evaluate_predictions_pro}
\alias{evaluate_predictions_pro}
\title{Evaluate External Predictions}
\usage{
evaluate_predictions_pro(prediction_df, years_to_evaluate = c(1, 3, 5))
}
\arguments{
\item{prediction_df}{Data frame with columns \code{time}, \code{outcome}, \code{score}, \code{ID}.}

\item{years_to_evaluate}{Years for AUC.}
}
\value{
List of evaluation metrics.
}
\description{
Calculates performance metrics for external prediction sets.
}
