% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis.R
\name{gbm_pro}
\alias{gbm_pro}
\title{Train Gradient Boosting Machine (GBM) for Survival}
\usage{
gbm_pro(X, y_surv, tune = FALSE, cv.folds = 5, max_tune_iter = 10)
}
\arguments{
\item{X}{A data frame of predictors.}

\item{y_surv}{A \code{Surv} object.}

\item{tune}{Logical. If TRUE, performs random search.}

\item{cv.folds}{Integer. Number of cross-validation folds.}

\item{max_tune_iter}{Integer. Maximum iterations for random search.}
}
\value{
An object of class \code{survival_gbm} and \code{pro_model}.
}
\description{
Fits a stochastic gradient boosting model using the Cox Partial Likelihood distribution.
Supports random search for hyperparameter optimization.
}
