% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis.R
\name{get_registered_models_pro}
\alias{get_registered_models_pro}
\title{Get Registered Prognostic Models}
\usage{
get_registered_models_pro()
}
\value{
Named list of functions.
}
\description{
Retrieves the list of available models.
}
