% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis.R
\name{initialize_modeling_system_pro}
\alias{initialize_modeling_system_pro}
\title{Initialize Prognosis Modeling System}
\usage{
initialize_modeling_system_pro()
}
\description{
Initializes the environment and registers default survival models
(Lasso, Elastic Net, Ridge, RSF, StepCox, GBM, XGBoost, PLS).
}
