% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis.R
\name{pls_pro}
\alias{pls_pro}
\title{Train Partial Least Squares Cox (PLS-Cox)}
\usage{
pls_pro(X, y_surv, tune = FALSE)
}
\arguments{
\item{X}{A data frame of predictors.}

\item{y_surv}{A \code{Surv} object containing time and status.}

\item{tune}{Logical. If TRUE, performs internal tuning (currently handled by cv.glmnet automatically).}
}
\value{
An object of class \code{survival_plsRcox} and \code{pro_model}.
}
\description{
Fits a Cox model using Partial Least Squares reduction for high-dimensional data.
}
