% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis.R
\name{print_model_summary_pro}
\alias{print_model_summary_pro}
\title{Print Prognostic Model Summary}
\usage{
print_model_summary_pro(model_name, results_list)
}
\arguments{
\item{model_name}{Name of the model.}

\item{results_list}{Result object containing \code{evaluation_metrics}.}
}
\description{
Formatted console output of model performance.
}
