% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis.R
\name{register_model_pro}
\alias{register_model_pro}
\title{Register a Prognostic Model}
\usage{
register_model_pro(name, func)
}
\arguments{
\item{name}{String identifier for the model.}

\item{func}{The model training function.}
}
\description{
Registers a model function into the internal system environment, making it available for batch execution.
}
