% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_dia}
\alias{test_dia}
\title{Test Data for Diagnostic Models}
\format{
A data frame with rows for samples and 22 columns:
\describe{
\item{sample}{character. Unique identifier for each sample.}
\item{outcome}{integer. The binary outcome (0 or 1).}
\item{AC004637.1}{numeric. Gene expression level.}
\item{AC008459.1}{numeric. Gene expression level.}
\item{AC009242.1}{numeric. Gene expression level.}
\item{AC016735.1}{numeric. Gene expression level.}
\item{AC090125.1}{numeric. Gene expression level.}
\item{AC104237.3}{numeric. Gene expression level.}
\item{AC112721.2}{numeric. Gene expression level.}
\item{AC246817.1}{numeric. Gene expression level.}
\item{AL135841.1}{numeric. Gene expression level.}
\item{AL139241.1}{numeric. Gene expression level.}
\item{HYMAI}{numeric. Gene expression level.}
\item{KCNIP2.AS1}{numeric. Gene expression level.}
\item{LINC00639}{numeric. Gene expression level.}
\item{LINC00922}{numeric. Gene expression level.}
\item{LINC00924}{numeric. Gene expression level.}
\item{LINC00958}{numeric. Gene expression level.}
\item{LINC01028}{numeric. Gene expression level.}
\item{LINC01614}{numeric. Gene expression level.}
\item{LINC01644}{numeric. Gene expression level.}
\item{PRDM16.DT}{numeric. Gene expression level.}
}
}
\source{
Stored in \code{data/test_dia.rda}.
}
\usage{
test_dia
}
\description{
A test dataset for evaluating diagnostic models, with a structure
identical to \code{train_dia}.
}
\keyword{datasets}
