% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_pro}
\alias{test_pro}
\title{Test Data for Prognostic (Survival) Models}
\format{
A data frame with rows for samples and 31 columns:
\describe{
\item{sample}{character. Unique identifier for each sample.}
\item{outcome}{integer. The event status (0 or 1).}
\item{time}{numeric. The time to event or censoring.}
\item{AC004990.1}{numeric. Gene expression level.}
\item{AC055854.1}{numeric. Gene expression level.}
\item{AC084212.1}{numeric. Gene expression level.}
\item{AC092118.1}{numeric. Gene expression level.}
\item{AC093515.1}{numeric. Gene expression level.}
\item{AC104211.1}{numeric. Gene expression level.}
\item{AC105046.1}{numeric. Gene expression level.}
\item{AC105219.1}{numeric. Gene expression level.}
\item{AC110772.2}{numeric. Gene expression level.}
\item{AC133644.1}{numeric. Gene expression level.}
\item{AL133467.1}{numeric. Gene expression level.}
\item{AL391845.2}{numeric. Gene expression level.}
\item{AL590434.1}{numeric. Gene expression level.}
\item{AL603840.1}{numeric. Gene expression level.}
\item{AP000851.2}{numeric. Gene expression level.}
\item{AP001434.1}{numeric. Gene expression level.}
\item{C9orf163}{numeric. Gene expression level.}
\item{FAM153CP}{numeric. Gene expression level.}
\item{HOTAIR}{numeric. Gene expression level.}
\item{HYMAI}{numeric. Gene expression level.}
\item{LINC00165}{numeric. Gene expression level.}
\item{LINC01028}{numeric. Gene expression level.}
\item{LINC01152}{numeric. Gene expression level.}
\item{LINC01497}{numeric. Gene expression level.}
\item{LINC01614}{numeric. Gene expression level.}
\item{LINC01929}{numeric. Gene expression level.}
\item{LINC02408}{numeric. Gene expression level.}
\item{SIRLNT}{numeric. Gene expression level.}
}
}
\source{
Stored in \code{data/test_pro.rda}.
}
\usage{
test_pro
}
\description{
A test dataset for evaluating prognostic models, with a structure
identical to \code{train_pro}.
}
\keyword{datasets}
