% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis.R
\name{xgb_pro}
\alias{xgb_pro}
\title{Train XGBoost Cox Model}
\usage{
xgb_pro(X, y_surv, tune = FALSE)
}
\arguments{
\item{X}{A data frame of predictors.}

\item{y_surv}{A \code{Surv} object containing time and status.}

\item{tune}{Logical. If TRUE, performs internal tuning (currently handled by cv.glmnet automatically).}
}
\value{
An object of class \code{survival_xgboost} and \code{pro_model}.
}
\description{
Fits an XGBoost model using the Cox proportional hazards objective function.
}
