% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis.R
\name{models_pro}
\alias{models_pro}
\title{Run Multiple Prognostic Models}
\usage{
models_pro(
  data,
  model = "all_pro",
  tune = FALSE,
  seed = 123,
  time_unit = "day",
  years_to_evaluate = c(1, 3, 5)
)
}
\arguments{
\item{data}{Input data frame.}

\item{model}{Character vector of model names or "all_pro".}

\item{tune}{Logical. Enable hyperparameter tuning?}

\item{seed}{Random seed.}

\item{time_unit}{Time unit of input.}

\item{years_to_evaluate}{Years for AUC calculation.}
}
\value{
A list of model results.
}
\description{
High-level API to train and evaluate multiple survival models in batch.
}
