% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prognosis.R
\name{rsf_pro}
\alias{rsf_pro}
\title{Train Random Survival Forest (RSF)}
\usage{
rsf_pro(X, y_surv, tune = FALSE, tune_params = NULL)
}
\arguments{
\item{X}{A data frame of predictors.}

\item{y_surv}{A \code{Surv} object containing time and status.}

\item{tune}{Logical. If TRUE, performs grid search for optimal hyperparameters based on C-index.}

\item{tune_params}{Optional data frame containing the grid for tuning.}
}
\value{
An object of class \code{survival_rsf} and \code{pro_model}.
}
\description{
Fits a Random Survival Forest using the log-rank splitting rule.
Includes capabilities for hyperparameter tuning via grid search over \code{ntree},
\code{nodesize}, and \code{mtry}.
}
