% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{xb_dia}
\alias{xb_dia}
\title{Train an XGBoost Tree Model for Classification}
\usage{
xb_dia(X, y, tune = FALSE, cv_folds = 5, tune_length = 20)
}
\arguments{
\item{X}{A data frame of features.}

\item{y}{A factor vector of class labels.}

\item{tune}{Logical, whether to perform hyperparameter tuning using \code{caret}'s
default grid (if \code{TRUE}) or use fixed values (if \code{FALSE}).}

\item{cv_folds}{An integer, the number of cross-validation folds for \code{caret}.}

\item{tune_length}{An integer, the number of random parameter combinations to try
when tune=TRUE. Only used when search="random". Default is 20.}
}
\value{
A \code{caret::train} object representing the trained XGBoost model.
}
\description{
Trains an Extreme Gradient Boosting (XGBoost) model using \code{caret::train}
for binary classification.
}
